+++ packages/twenty-docker/twenty-app-dev/Dockerfile
@@ -102,7 +102,7 @@ ENV PG_DATABASE_URL=postgres://twenty:twenty@localhost:5432/default \
     SERVER_URL=http://localhost:2020 \
     REDIS_URL=redis://localhost:6379 \
     STORAGE_TYPE=local \
-    APP_SECRET=change-me-in-production \
+    APP_SECRET=twenty-app-dev-secret-not-for-production \
     REACT_APP_SERVER_BASE_URL=$REACT_APP_SERVER_BASE_URL \
     APP_VERSION=$APP_VERSION \
     NODE_ENV=development \
